requirements_v2.txt
# Artificial Teacher v2.0 — Dependencies

# Telegram Bot
aiogram>=3.4
aiosqlite>=0.20

# AI & HTTP
httpx>=0.27
openai>=1.0

# Configuration
pydantic-settings>=2.0
python-dotenv>=1.0

# Template rendering
Jinja2>=3.1

# Scheduling
APScheduler>=3.10

# FastAPI
fastapi>=0.110
uvicorn[standard]>=0.29

# Testing
pytest>=8.0
pytest-asyncio>=0.23
